Fix automations detail route being shadowed by the list route
The list lived at _chat.automations.tsx, which made it the PARENT of
_chat.automations.$automationId.tsx. With no <Outlet/>, navigating to a detail
URL just re-rendered the list and the detail page was unreachable. Convert
_chat.automations.tsx into a layout (Outlet only) and move the list into
_chat.automations.index.tsx so /automations and /automations/$automationId each
render as their own full page. Verified live: the detail route now renders.
