+++ apps/web/src/routes/_chat.automations.index.tsx
@@ -1,0 +1,170 @@
+import { type AutomationDefinition, type AutomationRun } from "@t3tools/contracts";
+import { createFileRoute, useNavigate } from "@tanstack/react-router";
+import { useState } from "react";
+
+import { SidebarHeaderNavigationControls } from "~/components/SidebarHeaderNavigationControls";
+import { Button } from "~/components/ui/button";
+import { PlusIcon, RefreshCwIcon } from "~/lib/icons";
+import { cn } from "~/lib/utils";
+import { useStore } from "~/store";
+import {
+  type AutomationFormState,
+  AutomationDialog,
+  automationStatusDotClass,
+  createInputFromForm,
+  formatCadence,
+  formFromDefinition,
+  isFormSubmittable,
+  updateInputFromForm,
+  useAutomations,
+} from "./-automations.shared";
+import { resolveThreadPickerTitle } from "./-chatThreadRoute.logic";
+
+export const Route = createFileRoute("/_chat/automations/")({
+  component: AutomationsRouteView,
+});
+
+function AutomationsRouteView() {
+  const navigate = useNavigate();
+  const projects = useStore((state) => state.projects);
+  const threads = useStore((state) => state.threads);
+  const [editingDefinition, setEditingDefinition] = useState<AutomationDefinition | null>(null);
+  const [dialogOpen, setDialogOpen] = useState(false);
+  const fallbackProjectId = projects[0]?.id ?? "";
+  const [form, setForm] = useState<AutomationFormState>(() =>
+    formFromDefinition(null, fallbackProjectId),
+  );
+
+  const { data, isLoading, refetch, createMutation, updateMutation, runsByAutomationId } =
+    useAutomations((threadId) => void navigate({ to: "/$threadId", params: { threadId } }));
+
+  const openCreateDialog = () => {
+    setEditingDefinition(null);
+    setForm(formFromDefinition(null, fallbackProjectId));
+    setDialogOpen(true);
+  };
+
+  const submitForm = () => {
+    if (!isFormSubmittable(form)) return;
+    const closeOnSuccess = { onSuccess: () => setDialogOpen(false) };
+    if (editingDefinition) {
+      updateMutation.mutate(updateInputFromForm(editingDefinition, form, projects), closeOnSuccess);
+      return;
+    }
+    createMutation.mutate(createInputFromForm(form, projects), closeOnSuccess);
+  };
+
+  const active = data.definitions.filter((definition) => definition.enabled);
+  const paused = data.definitions.filter((definition) => !definition.enabled);
+
+  const projectName = (definition: AutomationDefinition) =>
+    projects.find((project) => project.id === definition.projectId)?.name ?? "Unknown project";
+
+  const subtitle = (definition: AutomationDefinition) => {
+    if (definition.mode === "heartbeat") {
+      const thread = threads.find((candidate) => candidate.id === definition.targetThreadId);
+      const target = thread ? resolveThreadPickerTitle(thread.title) : projectName(definition);
+      return `Heartbeat • ${target}`;
+    }
+    return projectName(definition);
+  };
+
+  const renderRow = (definition: AutomationDefinition) => {
+    const latestRun: AutomationRun | null = runsByAutomationId.get(definition.id)?.[0] ?? null;
+    return (
+      <button
+        key={definition.id}
+        type="button"
+        onClick={() =>
+          void navigate({
+            to: "/automations/$automationId",
+            params: { automationId: definition.id },
+          })
+        }
+        className="flex h-12 w-full items-center gap-3 border-b border-border/60 px-3 text-left transition-colors last:border-b-0 hover:bg-[var(--color-background-elevated-secondary)]"
+      >
+        <span className={cn("shrink-0", automationStatusDotClass(definition, latestRun))}>
+          <span className="block size-2 rounded-full bg-current" />
+        </span>
+        <div className="min-w-0 flex-1">
+          <div className="truncate text-sm font-medium text-foreground">{definition.name}</div>
+          <div className="truncate text-xs text-muted-foreground">{subtitle(definition)}</div>
+        </div>
+        <span className="shrink-0 text-xs text-muted-foreground">
+          {definition.enabled ? formatCadence(definition.schedule) : "Paused"}
+        </span>
+      </button>
+    );
+  };
+
+  return (
+    <div className="flex h-full min-h-0 flex-col bg-background text-foreground">
+      <header className="drag-region flex h-12 shrink-0 items-center gap-3 border-b border-border/60 px-3">
+        <SidebarHeaderNavigationControls />
+        <div className="min-w-0 flex-1">
+          <h1 className="truncate font-heading text-sm font-semibold">Automations</h1>
+        </div>
+        <div className="flex items-center gap-2">
+          <Button type="button" size="sm" variant="ghost" onClick={() => void refetch()}>
+            <RefreshCwIcon className="size-4" />
+            Refresh
+          </Button>
+          <Button
+            type="button"
+            size="sm"
+            onClick={openCreateDialog}
+            disabled={projects.length === 0}
+          >
+            <PlusIcon className="size-4" />
+            New
+          </Button>
+        </div>
+      </header>
+
+      <main className="min-h-0 flex-1 overflow-y-auto">
+        <div className="mx-auto flex w-full max-w-3xl flex-col gap-6 px-4 py-6">
+          {isLoading ? (
+            <div className="rounded-lg border border-border bg-muted/20 px-4 py-8 text-center text-sm text-muted-foreground">
+              Loading automations...
+            </div>
+          ) : data.definitions.length === 0 ? (
+            <div className="rounded-lg border border-border bg-muted/20 px-4 py-8 text-center text-sm text-muted-foreground">
+              No automations yet.
+            </div>
+          ) : (
+            <>
+              {active.length > 0 ? (
+                <section className="flex flex-col gap-1">
+                  <h2 className="px-3 text-xs font-medium text-muted-foreground">Current</h2>
+                  <div className="overflow-hidden rounded-lg border border-border">
+                    {active.map(renderRow)}
+                  </div>
+                </section>
+              ) : null}
+              {paused.length > 0 ? (
+                <section className="flex flex-col gap-1">
+                  <h2 className="px-3 text-xs font-medium text-muted-foreground">Paused</h2>
+                  <div className="overflow-hidden rounded-lg border border-border">
+                    {paused.map(renderRow)}
+                  </div>
+                </section>
+              ) : null}
+            </>
+          )}
+        </div>
+      </main>
+
+      <AutomationDialog
+        open={dialogOpen}
+        editing={editingDefinition !== null}
+        form={form}
+        projects={projects}
+        threads={threads}
+        onOpenChange={setDialogOpen}
+        onFormChange={setForm}
+        onSubmit={submitForm}
+        busy={createMutation.isPending || updateMutation.isPending}
+      />
+    </div>
+  );
+}
